requirements.txt
pandas
plotly-express
streamlit
nbformat